+++ src/App/App.js
@@ -1,0 +1,7 @@
+import React from "react";
+import Routes from "../routes/routes"
+const App = () => {
+  return <Routes />
+};
+
+export default App;
