+++ src/constants/paths.ts
@@ -1,0 +1,5 @@
+export const PATH = {
+    HOME: "/",
+    BUYDCUP: "/buy-dcup",
+  }
+
