+++ src/pages/Home/Home.tsx
@@ -1,0 +1,9 @@
+import React from "react"
+import MainLayout from "../../layouts/MainLayout"
+export default function Home() {
+  return (
+    <MainLayout>
+      <h2 className="mb-4">Home</h2>
+    </MainLayout>
+  )
+}
